backend/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
cffi==1.17.1
cssselect2==0.7.0
Django==5.1.4
django-cors-headers==4.6.0
et_xmlfile==2.0.0
fonttools==4.55.5
openpyxl==3.1.5
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
weasyprint==63.1
webencodings==0.5.1
zopfli==0.2.3.post1
